pyproject.toml
[project]
name = "agentic-maps"
version = "0.1.0"
description = "Agentic Maps"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.13.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
